Restore previous with new closing bracket
This commit sort of restores the previous setup and moves the closing
brackets/braces under the bottom row pinky.

This config should play better with a _regular_ muscle memory that
expects opening bracket/brace at the regular place with the only
specificity of the closing brackets/braces to be on the bottom row under
the pinky.

This frees the keys previously used for [{}] which can now hold pipe,
plus etc.